Add files via upload
I added GIFs to show what the animation should look like when hold frames are added properly. The sprite sheet will not include the hold frames and will look super fast without them. Whoever is adding the sprites, please be sure to add hold frames to the animations that need them, based on the provided GIFs. I didn't want to include hold frames on the sprite sheet because it wouldn't export properly.